+++ src/kernel/mm/pmm.h
@@ -19,7 +19,7 @@ enum { PAGE_SIZE = 4096, E820_USABLE = 1 };
 
 struct pmm_info {
     uint32_t page_size;
-    uint32_t managed_limit;
+    uint64_t managed_limit;
     uint32_t managed_pages;
     uint32_t free_pages;
     uint32_t used_pages;
