+++ src/kernel/syscall/sys_file.c
@@ -13,7 +13,7 @@ static int streq(const char *a, const char *b) {
     return *a == 0 && *b == 0;
 }
 
-int sys_open_console_aware(uint32_t path_arg, uint32_t flags, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_open_console_aware(uintptr_t path_arg, uintptr_t flags, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)c; (void)d; (void)e;
     const char *path = (const char *)(uintptr_t)path_arg;
     if (!user_string_ok(path))
@@ -23,83 +23,83 @@ int sys_open_console_aware(uint32_t path_arg, uint32_t flags, uint32_t c, uint32
     return vfs_open_flags(path, (int)flags);
 }
 
-int sys_close(uint32_t fd, uint32_t b, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_close(uintptr_t fd, uintptr_t b, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)b; (void)c; (void)d; (void)e;
     return vfs_close((int)fd);
 }
 
-int sys_read(uint32_t fd, uint32_t buf, uint32_t count, uint32_t d, uint32_t e) {
+intptr_t sys_read(uintptr_t fd, uintptr_t buf, uintptr_t count, uintptr_t d, uintptr_t e) {
     (void)d; (void)e;
     if (!user_range_writable(buf, count))
         return -1;
     return vfs_read((int)fd, (void *)(uintptr_t)buf, (size_t)count);
 }
 
-int sys_write(uint32_t fd, uint32_t buf, uint32_t count, uint32_t d, uint32_t e) {
+intptr_t sys_write(uintptr_t fd, uintptr_t buf, uintptr_t count, uintptr_t d, uintptr_t e) {
     (void)d; (void)e;
     if (!user_range_ok(buf, count))
         return -1;
     return vfs_write((int)fd, (const void *)(uintptr_t)buf, (size_t)count);
 }
 
-int sys_dup(uint32_t fd, uint32_t b, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_dup(uintptr_t fd, uintptr_t b, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)b; (void)c; (void)d; (void)e;
     return vfs_dup((int)fd);
 }
 
-int sys_dup2(uint32_t oldfd, uint32_t newfd, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_dup2(uintptr_t oldfd, uintptr_t newfd, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)c; (void)d; (void)e;
     return vfs_dup2((int)oldfd, (int)newfd);
 }
 
-int sys_stat(uint32_t path, uint32_t st, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_stat(uintptr_t path, uintptr_t st, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)c; (void)d; (void)e;
     if (!user_string_ok((const char *)(uintptr_t)path) || !user_range_writable(st, sizeof(struct stat)))
         return -1;
     return vfs_stat((const char *)(uintptr_t)path, (struct stat *)(uintptr_t)st);
 }
 
-int sys_getdents(uint32_t fd, uint32_t ents, uint32_t count, uint32_t d, uint32_t e) {
+intptr_t sys_getdents(uintptr_t fd, uintptr_t ents, uintptr_t count, uintptr_t d, uintptr_t e) {
     (void)d; (void)e;
     if (!user_range_writable(ents, count))
         return -1;
     return vfs_getdents((int)fd, (struct dirent *)(uintptr_t)ents, (size_t)count);
 }
 
-int sys_mkdir(uint32_t path, uint32_t b, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_mkdir(uintptr_t path, uintptr_t b, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)b; (void)c; (void)d; (void)e;
     if (!user_string_ok((const char *)(uintptr_t)path))
         return -1;
     return vfs_mkdir((const char *)(uintptr_t)path);
 }
 
-int sys_unlink(uint32_t path, uint32_t b, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_unlink(uintptr_t path, uintptr_t b, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)b; (void)c; (void)d; (void)e;
     if (!user_string_ok((const char *)(uintptr_t)path))
         return -1;
     return vfs_remove((const char *)(uintptr_t)path);
 }
 
-int sys_create(uint32_t path, uint32_t b, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_create(uintptr_t path, uintptr_t b, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)b; (void)c; (void)d; (void)e;
     if (!user_string_ok((const char *)(uintptr_t)path))
         return -1;
     return vfs_create((const char *)(uintptr_t)path);
 }
 
-int sys_lseek(uint32_t fd, uint32_t offset, uint32_t whence, uint32_t d, uint32_t e) {
+intptr_t sys_lseek(uintptr_t fd, uintptr_t offset, uintptr_t whence, uintptr_t d, uintptr_t e) {
     (void)d; (void)e;
     return vfs_lseek((int)fd, (int)offset, (int)whence);
 }
 
-int sys_rmdir(uint32_t path, uint32_t b, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_rmdir(uintptr_t path, uintptr_t b, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)b; (void)c; (void)d; (void)e;
     if (!user_string_ok((const char *)(uintptr_t)path))
         return -1;
     return vfs_rmdir((const char *)(uintptr_t)path);
 }
 
-int sys_rename(uint32_t old_path, uint32_t new_path, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_rename(uintptr_t old_path, uintptr_t new_path, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)c; (void)d; (void)e;
     if (!user_string_ok((const char *)(uintptr_t)old_path) ||
         !user_string_ok((const char *)(uintptr_t)new_path))
@@ -107,7 +107,7 @@ int sys_rename(uint32_t old_path, uint32_t new_path, uint32_t c, uint32_t d, uin
     return vfs_rename((const char *)(uintptr_t)old_path, (const char *)(uintptr_t)new_path);
 }
 
-int sys_fsstat(uint32_t info_arg, uint32_t b, uint32_t c, uint32_t d, uint32_t e) {
+intptr_t sys_fsstat(uintptr_t info_arg, uintptr_t b, uintptr_t c, uintptr_t d, uintptr_t e) {
     (void)b; (void)c; (void)d; (void)e;
     if (!user_range_writable(info_arg, sizeof(struct fs_info)))
         return -1;
